Add dirty mode for sync command
This will help debugging and allow user scripts to check properties
after sync.

+++ carbonate/sync.py
@@ -91,7 +91,7 @@ def heal_metric(source, dest):
             logging.warn("Failed to copy %s! %s" % (dest, e))
 
 
-def run_batch(metrics_to_sync, remote, local_storage, rsync_options):
+def run_batch(metrics_to_sync, remote, local_storage, rsync_options, dirty):
     staging_dir = mkdtemp(prefix=remote)
     sync_file = NamedTemporaryFile(delete=False)
 
@@ -123,5 +123,9 @@ def run_batch(metrics_to_sync, remote, local_storage, rsync_options):
     print "    Total time: %ss" % total_time
 
     # Cleanup
-    rmtree(staging_dir)
+    if dirty:
+      print "    dirty mode: left temporary directory %s" % staging_dir
+    else:
+      rmtree(staging_dir)
+
     os.unlink(sync_file.name)
